implemented uploading cv for about section

+++ app/Http/Controllers/API/About/DeleteController.php
@@ -13,9 +13,16 @@ class DeleteController extends BaseController
 {
     public function __invoke(About $about)
     {
+        
+        if($about->path_image) {
+            Storage::disk('public')->delete($about->path_image);
+            Storage::disk('public')->delete(str_replace('images/', 'images/prev_', $about->path_image ));
+        }
+        if($about->path_cv) {
+            Storage::disk('public')->delete($about->path_cv);
+        }
+       
         $about->delete();
-        Storage::disk('public')->delete($about->path_image);
-        Storage::disk('public')->delete(str_replace('images/', 'images/prev_', $about->path_image ));
         return response([]);
     }
 }
